Lets user choose profile
