+++ fafi/cli.py
@@ -1,0 +1,11 @@
+def let_user_pick(options):
+    print("Please choose:")
+    for idx, element in enumerate(options):
+        print("{}) {}".format(idx + 1, element))
+    i = input("Enter number: ")
+    try:
+        if 0 < int(i) <= len(options):
+            return int(i)
+    except:
+        pass
+    return None
