+++ pyproject.toml
@@ -13,7 +13,7 @@ click = "^7.1.1"
 appdirs = "^1.4.3"
 
 [tool.poetry.dev-dependencies]
-black = "^19.10b0"
+black = "^20.8b1"
 pylint = "^2.6.0"
 
 [tool.poetry.scripts]
